+++ src/App.jsx
@@ -1,21 +1,21 @@
 import React from "react";
-import Menu from "./common/Menu.jsx"
+import MenuLayout from "./components/common/layout/MenuLayout.jsx";
 import { Route, Switch } from "react-router-dom";
 import "antd/dist/antd.css";
-import "./index.css"
-import { PersonalData } from "./components/PersonalData.jsx";
+import "./index.less"
 import { Provider } from "react-redux";
 import store from "./redux/store.js";
-import ParentForm from "./components/ParentForm.jsx";
+import SignupForm from "./container/SignupForm.jsx";
 
 function App() {
     return (
         <>
             <Provider store={store}>
-                <Menu />
-                <Switch>
-                    <Route exact path='/signup' component={() => <><ParentForm /></>} />
-                </Switch>
+                <MenuLayout className="layout-main">
+                    <Switch>
+                        <Route exact path='/signup' component={() => <><SignupForm /></>} />
+                    </Switch>
+                </MenuLayout>
             </Provider>
         </>
     );
