+++ src/components/family/FamilyMember.jsx
@@ -1,11 +1,8 @@
 import React, { useState } from 'react';
 import { Typography } from 'antd';
 import { useSelector } from 'react-redux';
-import { Row, Col } from 'antd';
 import { Form, Input, Button, Space, DatePicker, Select, Checkbox } from 'antd';
-import InsuranceDetail from './InsuranceDetail.jsx';
-import CheckboxInsurance from './CheckboxInsurance.jsx';
-import FormLayout from './FormLayout.jsx';
+import InsuranceDetail from '../insurance/InsuranceDetail.jsx';
 const { Title } = Typography;
 
 const FamilyMember = () => {
@@ -17,28 +14,22 @@ const FamilyMember = () => {
 
     return (
         <>
-            {/* <FormLayout> */}
-            {/* {fields.map(({ key, name, fieldKey, ...restField }) => ( */}
             <Form name="dynamic_form_nest_item" autoComplete="on">
-                {/* <Form.List name="users"> */}
                 <Space style={{ display: 'flex', marginBottom: 8 }} align="baseline">
                     <Form.Item
-                        // {...restField}
-                        name={[name, 'first']}
+                        name={['first']}
                         rules={[{ required: true, message: 'Missing first name' }]}
                     >
                         <Input placeholder="First Name" />
                     </Form.Item>
                     <Form.Item
-                        // {...restField}
-                        name={[name, 'last']}
+                        name={['last']}
                         rules={[{ required: true, message: 'Missing last name' }]}
                     >
                         <Input placeholder="Last Name" />
                     </Form.Item>
                     <Form.Item
-                        // {...restField}
-                        name={[name, "date-picker"]}
+                        name={["date-picker"]}
                         rules={[
                             {
                                 type: 'object',
@@ -49,8 +40,7 @@ const FamilyMember = () => {
                         <DatePicker />
                     </Form.Item>
                     <Form.Item
-                        // {...restField}
-                        name={[name, "insurancestatus"]}
+                        name={["insurancestatus"]}
                         rules={[
                             {
                                 required: true,
@@ -65,23 +55,14 @@ const FamilyMember = () => {
                         </Select>
                     </Form.Item>
 
-                    <Form.Item
-                        name={[name, 'details']}
-                        rules={[{ required: true, message: 'Missing status details' }]}
-                    >
-
-                        {status === "other" ?
-                            <InsuranceDetail /> : status === "none" ?
-                                (
-                                    <Checkbox>Attest Insurance</Checkbox>
-                                ) :
-                                null}
-                    </Form.Item>
+                    {status === "other" ?
+                        <InsuranceDetail /> : status === "none" ?
+                            (
+                                <Checkbox>Attest Insurance</Checkbox>
+                            ) :
+                            null}
                 </Space>
-                {/* </Form.List> */}
             </Form>
-            {/* ))} */}
-            {/* </FormLayout> */}
         </>
     )
 }
