--- src/index.css
@@ -1,204 +1,0 @@
-body {
-    margin: 0;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-      sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    overflow-x: hidden;
-  }
-  
-  code {
-    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-      monospace;
-  }
-  
-  .main_div {
-    width: 100%;
-    height: 120vh;
-    background: #6983aa;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-  }
-  
-  .center_div {
-    width: 50%;
-    height: 110vh;
-    background-color: white;
-    box-shadow: 5px 5px 25px -5px rgba(0, 0, 0, 0.5);
-    border-radius: 15px;
-
-  }
-
-  .switch {
-    width: 100%;
-    height: 10vh;
-    background: #6983aa;
-    background-color: white;
-    justify-content: flex-start;
-    align-items: flex-start;
-    text-align: start;
-  }
-
-  .fam_div {
-    width: 200%;
-    height: 80vh;
-    background-color: white;
-    border-radius: 5px;
-    margin-left: 10PX;
-  }
-
-  #h3 {
-    text-align: center;
-  }
-  .ant-form-item-control-input {
-    text-align: left;
-  }
-
-  .heading_style {
-    padding: 30px 0;
-    text-align: center;
-    font-family: Arial, Helvetica, sans-serif;
-    text-transform: uppercase;
-    font-size: 13px;
-    letter-spacing: 2px;
-    font-weight: 500;
-    color:#868686;
-    background-color: rgb(233, 232, 232);
-    font-size: 1.5rem;
-    margin-bottom: 30px;
-  }
-  
-  .usernotfound {
-    padding: 30px 0;
-    text-align: center;
-    font-family: Arial, Helvetica, sans-serif;
-    text-transform: uppercase;
-    font-size: 13px;
-    letter-spacing: 2px;
-    font-weight: 500;
-    color:#868686;
-    background-color: white;
-    font-size: 1.5rem;
-    margin-bottom: 30px;
-  }
-  .cards {
-    width: 100%;
-    height: auto;
-  }
-  
-  .card {
-    margin: 6%;
-    transition: all 0.4s cubic-bezier(0.175, 0.885, 0, 1);
-    background-color: rgb(233, 232, 232);
-    width: 21.25%;
-    border-radius: 12px 12px 12px 12px;
-    box-shadow: 0px 13px 10px -7px rgba(0, 0, 0, 0.1);
-    float: left;
-  }
-  
-  .User {
-    margin: 6%;
-    transition: all 0.4s cubic-bezier(0.175, 0.885, 0, 1);
-    background-color: rgb(233, 232, 232);
-    width: 21.25%;
-    border-radius: 12px 12px 12px 12px;
-    box-shadow: 0px 13px 10px -7px rgba(0, 0, 0, 0.1);
-    align-self: center;
-  }
-  
-  .card:hover {
-    box-shadow: 0px 30px 18px -8px rgba(0, 0, 0, 0.1);
-    transform: scale(1.05, 1.05);
-  }
-  
-  .card_info {
-    font-family: Georgia, 'Times New Roman', Times, serif;
-    text-transform: uppercase;
-    font-size: 13px;
-    letter-spacing: 2px;
-    font-weight: 500;
-    color: #868686;
-    background-color: #fff;
-  }
-  
-  .card_img {
-    margin: auto;
-    padding: 20px 0 0 0;
-    display: block;
-  }
-  
-  .card_title {
-    background-color: #fff;
-    padding: 16px 0px 0px 24px;
-  }
-  
-  .card_text {
-    background-color: #fff;
-    padding: 6px 4px 24px 24px;
-  }
-  
-  button {
-    padding: 5px 8px;
-    font-family:Georgia, 'Times New Roman', Times, serif;
-    text-transform: uppercase;
-    font-size: 13px;
-    letter-spacing: 2px;
-    font-weight: 500;
-    color: #868686;
-    background-color: #fff;
-    outline: none;
-    border: 1px solid black;
-    cursor: pointer;
-    padding: 16px 24px 24px 24px;
-    margin-left: 24px;
-    margin-bottom: 16px;
-  }
-  
-  .searchbar {
-    width: 100%;
-    height: auto;
-    margin: 0 auto;
-    text-align: center;
-  }
-  
-  .searchbar input {
-    margin: 30px 0;
-    padding: 5px 10px;
-    border-radius: 5px;
-    font-family:Verdana, Geneva, Tahoma, sans-serif;
-  }
-  
-  a {
-    color: aqua;
-  }
-  
-  a.active_class {
-    color: white;
-  }
-
-  .menu_style {
-    width: 100%;
-    line-height: 60px;
-    background-color: #20232a;
-    display: flex;
-    justify-content: space-around;
-  
-  }
-
-  .continueButton {
-    position:relative;
-    bottom: 10px;
-    right:0%;
-  }
-
-  .ant-btn-block {
-    width: 50%;
-}
-
-h3.ant-typography {
-  text-align: center;
-}
