+++ src/index.less
@@ -1,0 +1,52 @@
+body {
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+      sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    overflow-x: hidden;
+  }
+  
+  code {
+    font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+      monospace;
+  }
+
+  #h3 {
+    text-align: center;
+  }
+  .ant-form-item-control-input {
+    text-align: left;
+  }
+  
+  a {
+    color: aqua;
+  }
+  
+  a.active_class {
+    color: white;
+  }
+
+  .menu_style {
+    width: 100%;
+    line-height: 60px;
+    background-color: #20232a;
+    display: flex;
+    justify-content: space-around;
+  
+  }
+
+  .continueButton {
+    position:relative;
+    bottom: 10px;
+    right:0%;
+  }
+
+  .ant-btn-block {
+    width: 50%;
+}
+
+h3.ant-typography {
+  text-align: center;
+}
